+++ dflib-echarts/src/main/java/org/dflib/echarts/LineItemStyle.java
@@ -7,7 +7,7 @@ import org.dflib.echarts.render.option.series.ItemStyleModel;
  */
 public class LineItemStyle {
 
-    ValOrColumn<String> color;
+    ValOrSeries<String> color;
     private Double opacity;
 
     public static LineItemStyle of() {
@@ -15,7 +15,7 @@ public class LineItemStyle {
     }
 
     public LineItemStyle color(String color) {
-        this.color = ValOrColumn.ofVal(color);
+        this.color = ValOrSeries.ofVal(color);
         return this;
     }
 
@@ -26,7 +26,7 @@ public class LineItemStyle {
      * @since 2.0.0
      */
     public LineItemStyle colorData(String colorDataColumn) {
-        this.color = ValOrColumn.ofDataColumn(colorDataColumn);
+        this.color = ValOrSeries.ofSeries(colorDataColumn);
         return this;
     }
 
@@ -38,7 +38,7 @@ public class LineItemStyle {
     ItemStyleModel resolve(Integer symbolSizeColorDimension) {
 
         String color = this.color != null
-                ? (this.color.val != null ? "'" + this.color.val + "'" : ValOrColumn.jsFunctionWithObjectParam(symbolSizeColorDimension))
+                ? (this.color.isVal() ? this.color.valQuotedString() : ValOrSeries.jsFunctionWithObjectParam(symbolSizeColorDimension))
                 : null;
 
         return new ItemStyleModel(